netbird/pyproject.toml
[project]
name = "netbird-manager"
version = "0.1.0"
description = "NetBird Groups and Peers Management Tool"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[project.scripts]
netbird-manager = "manage_groups:main"

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "freezegun>=1.5.5",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "respx>=0.22.0",
]

[tool.pytest.ini_options]
pythonpath = ["pynetbird/src"]
testpaths = ["pynetbird/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--cov=pynetbird",
    "--cov-branch",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
    "real_api: marks tests that require real API credentials",
]
